requirements.txt
# Core Deep Learning
tensorflow>=2.15.0
keras>=2.15.0

# Scientific Computing
numpy>=1.24.3
pandas>=2.0.3
scipy>=1.11.1

# Machine Learning
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.15.0

# Data Handling
openpyxl>=3.1.2
xlrd>=2.0.1

# Progress Bars
tqdm>=4.65.0

# Google Generative AI (Optional - for Gemini integration)
google-generativeai>=0.3.0

# Jupyter Support
ipykernel>=6.25.0
jupyterlab>=4.0.5

# Additional Utilities
python-dotenv>=1.0.0
requests>=2.31.0